Chat/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
markupsafe==2.1.3
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0
